flowrun/pyproject.toml
[project]
name = "flowrun"
version = "0.1.0"
description = "Simple YAML-driven workflow runner for automation tasks"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Your Name Here", email = "placeholder@example.com"}]
dependencies = [
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "httpx>=0.27.0",
]

[project.scripts]
flowrun = "flowrun.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/flowrun"]